# Improve unsupported coercion performance by optionally caching error