--- a/coercible.gemspec
+++ b/coercible.gemspec
@@ -23,4 +23,5 @@ Gem::Specification.new do |gem|
   gem.add_development_dependency 'rake', '~> 10.0'
   gem.add_development_dependency 'rspec', '~> 3.1.0'
   gem.add_development_dependency 'rspec-its', '~> 1.0'
+  gem.add_development_dependency 'benchmark-ips', '~> 2.7'
 end